Restructuring /fabric/core/rest directory
Moving api.go and api_test.go into the /fabric/core/rest directory.

Fixes #954
