+++ core/rest/api.go
@@ -17,7 +17,7 @@ specific language governing permissions and limitations
 under the License.
 */
 
-package core
+package rest
 
 import (
 	"errors"
